move tests to namespaces
