+++ tests/MudBlazor.Markdown.Tests/UrlOverrideTests/UrlOverrideTestsBase.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MudBlazor.Markdown.Tests.UrlOverrideTests
+{
+	public abstract class UrlOverrideTestsBase
+	{
+	}
+}
